defenseの行動を別ファイルに切り出した（途中
