+++ consai2_game/scripts/example/actions/center_back.py
@@ -1,17 +1,14 @@
 # coding: UTF-8
-# defense.pyでは、ボールを蹴らないactionを定義する
 
 import math
 import rospy
 import sys,os
 
-from consai2_msgs.msg import BallInfo, RobotInfo
-from consai2_msgs.msg import ControlTarget
+from consai2_msgs.msg import BallInfo, RobotInfo, ControlTarget
 from geometry_msgs.msg import Pose2D
 
 sys.path.append(os.pardir)
 from field import Field
-from observer import Observer
 import role
 import tool
 
@@ -166,234 +163,3 @@ def center_back(my_pose, ball_info, control_target, my_role, defense_num):
     control_target.path.append(target_pose)
 
     return control_target
-
-def _get_avoid_ball_pose(ball_pose, target_pose):
-    # target_poseがボールの半径0.5 m以内にある場合、
-    # ボールから離れたtarget_poseを生成する
-
-    THRESHOLD_DIST = 0.5 # meters
-    AVOID_DIST = 0.6 # meters
-
-    avoid_target_pose = target_pose
-
-    if tool.distance_2_poses(ball_pose, target_pose) < THRESHOLD_DIST:
-        angle_to_target = tool.get_angle(ball_pose, target_pose)
-        trans = tool.Trans(ball_pose, angle_to_target)
-        avoid_target_pose = trans.inverted_transform(Pose2D(AVOID_DIST, 0, 0))
-        # 目標角度を再設定
-        avoid_target_pose.theta = target_pose.theta
-
-    return avoid_target_pose
-    
-
-# ゾーンディフェンス
-def defense_zone(my_pose, ball_info, control_target, my_role, defense_num, their_robot_info, zone_enable):
-    # ゴールディフェンスに割り当てる台数
-    GOAL_DEFENSE_NUM = 2
-    # 現在のディフェンス数 - ゴールディフェンス数 = ゾーンディフェンスに割り当てられる台数
-    ZONE_DEFENSE_NUM = defense_num - GOAL_DEFENSE_NUM
-    # ゾーンディフェンスが始まるROLE_ID
-    ZONE_START_ROLE_ID = role.ROLE_ID["ROLE_ZONE_1"]
-    # ゾーンオフェンス用の待機場所
-    ZONE_OFFENCE_POSE = Pose2D(Field.field('length')*0.25 ,0,0)
-
-    # センターライン用のマージン
-    MARGIN_CENTER = 0.6
-    # ちょっと前進用のマージン
-    MARGIN_LITTLE_FORWARD = 1.0
-
-    # ドリブルパワー
-    DRIBBLE_POWER = 0.6
-
-    # ボール位置
-    ball_pose = ball_info.pose
-
-    # Field情報からペナルティエリアの情報を取得
-    # フィールド幅
-    field_width = Field.field('width')
-    # フィールド幅の半分
-    half_field_width = float(field_width) / 2
-    # フィールド幅の1/4
-    quarter_field_width = float(field_width) / 4
-    # フィールド長さ
-    field_length = Field.field('length')
-    # フィールド長さの1/4 → 自陣側の長さの半分
-    half_our_field_length = -float(field_length) / 4
-    # ゴール中心
-    goal_center = Field.goal_pose('our', 'center')
-
-    # ペナルティエリアの角
-    left_penalty_corner = Field.penalty_pose('our', 'upper_front')
-    right_penalty_corner = Field.penalty_pose('our', 'lower_front')
-
-    # 自分からボールへの角度（ボールの方向を向くため）
-    angle_to_ball = tool.get_angle(my_pose, ball_pose)
-    # ゴール中心からボールへの角度
-    angle_to_ball_from_goal = tool.get_angle(goal_center, ball_pose)
-
-    # ゾーンディフェンス用のID
-    zone_id = None
-    # 移動目標点の初期化
-    target_pose = Pose2D()
-
-    # ---------------------------------------------------------
-    # キックとドリブルはOFF
-    control_target.kick_power = 0.0
-    control_target.dribble_power = 0.0
-
-    # ゾーンオフェンス判定用フラグ
-    my_role_is_offence = False
-
-    # ボールが相手フィールドにあるとき
-    # ゾーンから1台ゾーンオフェンスに出す
-    # 相手キックオフ時などに前に出ないように
-    # マージンを持って相手フィールド側かを判断している
-    if ZONE_DEFENSE_NUM > 1 and ball_pose.x > MARGIN_CENTER:
-        # 1台ディフェンスが減る
-        ZONE_DEFENSE_NUM -= 1
-        # ゾーンディフェンスが始まるROLE_IDをずらす
-        ZONE_START_ROLE_ID = role.ROLE_ID["ROLE_ZONE_2"]
-        # ROLE_ZONE_1をゾーンオフェンスとして出す
-        if my_role is role.ROLE_ID["ROLE_ZONE_1"]:
-            my_role_is_offence = True
-
-    # 私はゾーンオフェンスです
-    if my_role_is_offence:
-        zone_id = 0
-        target_pose = ZONE_OFFENCE_POSE
-        # 基本的にアタッカーがボールを取りに行くので
-        # ボールが無い方向に移動してこぼれ球が取れるようにする
-        if ball_pose.y > 0:
-            target_pose.y  =  - quarter_field_width
-        else:
-            target_pose.y = quarter_field_width
-        # ボールを向く
-        target_pose.theta = angle_to_ball
-
-    # ゾーンオフェンス以外
-    if ZONE_DEFENSE_NUM > 0 and not my_role_is_offence:
-        step = float(field_width) / (ZONE_DEFENSE_NUM * 2)
-        # ゾーンディフェンスの数でフィールド幅を等分した配列を作る
-        split_field = [i * step - half_field_width for i in range(0,(ZONE_DEFENSE_NUM * 2 + 1))]
-        # ゾーンディフェンスの数でフィールド幅を等分した時の
-        # それぞれのゾーンの中心の配列を作る
-        split_field_center = [i * step - half_field_width for i in range(0,(ZONE_DEFENSE_NUM * 2)) \
-                if i % 2 != 0]
-
-        # 参照エラー対策のtry
-        try:
-            # ゾーンIDの計算
-            # ロボットが8台生きていてゾーンオフェンスがいなければ
-            # ゾーンIDは0~3
-            zone_id = my_role - ZONE_START_ROLE_ID
-            # ゾーンの中心を目標位置とする
-            target_pose.y = split_field_center[zone_id]
-
-            # 自分のゾーンに入っている敵チェック
-            # their_robot_infoのposeを抜き出してそれぞれチェック
-            # 敵が自陣側にいる、かつ、自分のゾーンの幅の中にいる、を確認
-            # 当てはまらない場合配列は空っぽ
-            invader_pose = [i.pose for i in their_robot_info \
-                    if split_field[zone_id * 2] < i.pose.y < split_field[(zone_id + 1) * 2] and \
-                    i.pose.x < 0]
-
-            # ボールが自分のゾーンの中に入っている, かつzone_enable
-            if(zone_enable and \
-                    ball_pose.x < 0 and \
-                    split_field[zone_id * 2] < ball_pose.y < split_field[(zone_id + 1) * 2]):
-                trans = tool.Trans(ball_pose, angle_to_ball_from_goal)
-                target_pose = trans.inverted_transform(Pose2D(-0.9, 0, 0))
-            # 自分のゾーンにボールはないけど敵がいる場合は割り込む
-            elif zone_enable and invader_pose != []:
-                # 敵とボールの間に割り込む
-                angle_to_ball_from_invader = tool.get_angle(invader_pose[0], ball_pose)
-                trans = tool.Trans(invader_pose[0], angle_to_ball_from_invader)
-                target_pose = trans.inverted_transform(Pose2D(0.5, 0, 0))
-            else:
-                # ボールが敵陣の時はディフェンスちょっと前進
-                if ball_pose.x > MARGIN_CENTER:
-                    target_pose.x = half_our_field_length + MARGIN_LITTLE_FORWARD
-                else:
-                    target_pose.x = half_our_field_length
-
-        except IndexError:
-            target_pose = my_pose
-        target_pose.theta = angle_to_ball
-
-    # ボールが来てたらボールを受け取る
-    if zone_id != None:
-        receive_ball_result, receive_target_pose = update_receive_ball(ball_info, my_pose, zone_id)
-        if receive_ball_result:
-            # ドリブラー回す
-            control_target.dribble_power = DRIBBLE_POWER
-            target_pose = receive_target_pose
-        else:
-            # ボールに近づいてたら離れる
-            target_pose = _get_avoid_ball_pose(ball_pose, target_pose)
-
-    # ペナルティエリアには入らない
-    if((left_penalty_corner.y + 0.2 > target_pose.y > right_penalty_corner.y - 0.2) and \
-                target_pose.x < left_penalty_corner.x + 0.3):
-        target_pose.x = half_our_field_length
-
-    control_target.path = []
-    control_target.path.append(target_pose)
-
-    return control_target
-
-# ボールレシーブ情報保持用のクラス
-class Receiving(object):
-    _recenving = [False] * role.ZONE_DEFENSE_NUM
-
-    @classmethod
-    def update_receiving(cls, zone_id, param):
-        Receiving._recenving[zone_id] = param
-    @classmethod
-    def receiving(cls, zone_id):
-        return Receiving._recenving[zone_id]
-
-
-def update_receive_ball(ball_info, my_pose, zone_id):
-    # ボール位置
-    ball_pose = ball_info.pose
-    # ボールスピード
-    ball_vel = ball_info.velocity
-
-    # 受け取れると判断する距離
-    _can_receive_dist = 1.0
-    # ヒステリシス
-    _can_receive_hysteresis = 0.3
-
-    result = False
-    target_pose = Pose2D()
-
-    # ボールが動いている
-    if Observer.ball_is_moving():
-        # ボール速度ベクトルの角度
-        angle_velocity = tool.get_angle_from_center(ball_vel)
-        trans = tool.Trans(ball_pose, angle_velocity)
-
-        tr_pose = trans.transform(my_pose)
-        
-        # ボール速度の線と垂直な距離
-        fabs_y = math.fabs(tr_pose.y)
-
-        # 受け取れる判定
-        if Receiving.receiving(zone_id) == False and \
-                fabs_y < _can_receive_dist - _can_receive_hysteresis:
-            Receiving.update_receiving(zone_id, True)
-        # 受け取れない判定
-        elif Receiving.receiving(zone_id) == True and \
-                fabs_y > _can_receive_dist + _can_receive_hysteresis:
-            Receiving.update_receiving(zone_id, False)
-
-        # 受け取れるかつボールが向かう方向にいる
-        if Receiving.receiving(zone_id) and tr_pose.x > 0.0:
-            tr_pose.y = 0.0
-            inv_pose = trans.inverted_transform(tr_pose)
-            angle_to_ball = tool.get_angle(inv_pose, ball_pose)
-            target_pose = Pose2D(inv_pose.x, inv_pose.y, angle_to_ball)
-            result = True
-
-    return result, target_pose
-
